Firmware:  - Lots of testing with ADC using smart Algorithms - nothing really worked so far  - Should be stable now
GUI:
 - Added a bit fanciness
 - Added release build version (Debug buttons are disabled)
 - Added debug auto test to change PWM frequency and ADC frequency

+++ Firmware/PCB/MIDI_Lighter/src/ADC.h
@@ -11,8 +11,8 @@
 
 #include <avr/io.h>
 
-void		ADC_Init(void);
-uint16_t	ADC_Get_Value(void);
+void			ADC_Init(void);
+uint32_t	ADC_Get_Value(void);
 
 
 #endif /* ADC_H_ */
